# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nexa Panel HTTP Interface
  version: 0.1.0
  description: Initial control-plane contract for the Nexa Panel foundation.
servers:
  - url: /api/v1
paths:
...
  /node/probe/observe:
    get:
      operationId: observeManagedProbe
      tags: [Node Operations]
      x-permission: system.read
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current state of the fixed Nexa-owned probe target.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeSnapshot"
        "503":
          description: The privileged node agent is unavailable.
  /node/probe/plan:
    post:
      operationId: planManagedProbe
      tags: [Node Operations]
      x-permission: operations.plan
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbeChange"
      responses:
        "200":
          description: A signed, time-limited plan issued by the privileged agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePlan"
  /node/probe/apply:
    post:
      operationId: applyManagedProbePlan
      tags: [Node Operations]
      x-permission: operations.apply
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodePlan"
      responses:
        "202":
          description: The signed plan was persisted as a durable job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /node/probe/rollback:
    post:
      operationId: rollbackManagedProbePlan
      tags: [Node Operations]
      x-permission: operations.apply
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodePlan"
      responses:
        "202":
          description: A rollback of the signed plan was persisted as a durable job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /postgresql/instances:
    get:
      operationId: listPostgreSQLInstances
      tags: [Databases]
      x-permission: databases.read
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Discovered PostgreSQL 16-18 instances, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: "#/components/schemas/PostgreSQLInstance"}}}}}}}
    post:
      operationId: createPostgreSQLInstance
      tags: [Databases]
      x-permission: databases.write
      security: [{sessionCookie: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreatePostgreSQLInstance"}}}}
      responses:
        "202": {description: PostgreSQL instance provisioning plan was queued}
...
  /admin-tools:
    get:
      operationId: listAdminTools
      tags: [Admin Tools]
      x-permission: databases.read
      security: [{sessionCookie: []}]
      responses: {"200": {description: Observed phpMyAdmin and pgAdmin deployment state}}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexa_session
...
  schemas:
...
    ProbeChange:
      type: object
      additionalProperties: false
      required: [present]
      properties:
        present:
          type: boolean
        content:
          type: string
          maxLength: 256
    NodeSnapshot:
      type: object
      required: [exists]
      properties:
        exists:
          type: boolean
        digest:
          type: string
          pattern: "^[a-f0-9]{64}$"
        content:
          type: string
          maxLength: 256
    NodePlan:
      type: object
      required: [id, kind, target, action, changed, before, desired, plannedAt, expiresAt, signature]
      properties:
        id:
          type: string
        kind:
          type: string
          const: nexa.probe.v1
        target:
          type: string
        action:
          type: string
          enum: [none, create, update, remove]
        changed:
          type: boolean
        before:
          $ref: "#/components/schemas/NodeSnapshot"
        desired:
          $ref: "#/components/schemas/NodeSnapshot"
        plannedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        signature:
          type: string
...
    PostgreSQLStatus:
      type: string
      enum: [planning, plan_ready, applying, active, online, down, backing_up, verified, restoring, failed]
    CreatePostgreSQLInstance:
      type: object
      additionalProperties: false
      required: [version, cluster]
      properties:
        version: {type: string, enum: ["16", "17", "18"]}
        cluster: {type: string, pattern: "^[a-z][a-z0-9_]{1,30}$"}
        port: {type: integer, minimum: 1024, maximum: 65535}
    PostgreSQLInstance:
      type: object
      required: [id, version, cluster, port, status, owner, dataPath, socketPath, logPath, configPath, systemdUnit, managedByNexa]
      properties:
        id: {type: string}
        version: {type: string, enum: ["16", "17", "18"]}
        cluster: {type: string}
        port: {type: integer}
        status: {$ref: "#/components/schemas/PostgreSQLStatus"}
        owner: {type: string}
        dataPath: {type: string}
        socketPath: {type: string}
        logPath: {type: string}
        configPath: {type: string}
        systemdUnit: {type: string}
        managedByNexa: {type: boolean}
        lastJobId: {type: integer, format: int64}
        failure: {type: string}
...
    Job:
      type: object
      required: [id, kind, state, progress, request, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        kind:
          type: string
        state:
          type: string
          enum: [queued, running, succeeded, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        actorUserId:
          type: string
        request:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
        failure:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...